# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  description: "# Summary\n  Given set of APIs for Bill management process\n# Versioning\n\
    \  By default, all requests receive the latest version of the REST API.  We encourage\
    \ you to explicitly request this version via the Accept header. Example \n   \
    \          \n  \"Accept: application/vnd.worldpay.bill-v1+json\"\n  \n# Security\
    \ \nAPIs use JWT Based Bearer Authentication. JWT Token must contain claims to\
    \ merchant and terminal operator identifiers at minimum.\n\n# Response Codes \n\
    \nA list of the  common HTTP response codes used in POS API is given below:\n\n\
    \  |HTTP Code|HTTP Verb|Description|\n  |--- |--- |--- |\n  |200|GET POST PUT|The\
    \ request has succeeded.|\n  |201|POST|The request has succeeded and a new resource\
    \ has been created as a result.|\n  |202|GET POST|The request has succeeded, but\
    \ the response body is incomplete.|\n  |204|PUT DELETE|The request has succeeded,\
    \ an existing resource has been updated/deleted and the response body is empty.|\n\
    \  |400|POST PUT|The service could not understand the request.|\n  |401|Any|The\
    \ client must authenticate itself.|\n  |404|GET PUT DELETE|The service can not\
    \ find requested resource.|\n  |409|POST|The service has experienced a conflict\
    \ between the data provided in the request and the data on file.|\n  |415|Any|Content-Type\
...
    \ within the service which hasn’t been accounted for.|\n  |503|Any|The service\
...
servers:
...
security:
- bearerAuth: []
...
paths:
  /bills:
    get:
      description: |
        Default all pending status bills be returned. Multiple status values can be provided with comma separated strings
      operationId: listBills
      parameters:
      - description: "All possible values for bill.\n\n`ready` - used to get the bill\
          \ from order management \n\n`pending` - once bill is open to be processed\n\
          \n`in-progress` - used to update once bill is picked-up for payment.\n\n\
          `complete` - once bill is completely paid\n\n`partial-complete` - if bill\
          \ is partially paid(in split payment) and wants to pay to the counter OR\
          \ customer wants to update the bill in un-happy path\n\n`cancel` - to cancel\
          \ a bill\n"
        explode: true
        in: query
        name: billStatus
        required: false
        schema:
          items:
            $ref: '#/components/schemas/listBills_billStatus_parameter_inner'
          type: array
        style: form
      - description: Bill tag of bill to be provided
        explode: true
        in: query
        name: billTag
        required: false
        schema:
          $ref: '#/components/schemas/billTagReference'
        style: form
      - explode: false
        in: header
        name: X-WP-User-Id
        required: false
        schema:
          $ref: '#/components/schemas/userReference'
        style: simple
      responses:
        "200":
          content:
            application/vnd.worldpay.bill-v1+json:
              schema:
                items:
                  $ref: '#/components/schemas/billDetails'
                type: array
          description: successful operation
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
          description: Invalid Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
          description: Resource Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
          description: Internal Server Error
      summary: List Bills by status and/or billTag
      tags:
      - Bill
    post:
      operationId: createBill
      parameters:
      - explode: false
        in: header
        name: X-WP-User-Id
        required: false
        schema:
          $ref: '#/components/schemas/userReference'
        style: simple
      requestBody:
        content:
          application/vnd.worldpay.bill-v1+json:
            schema:
              $ref: '#/components/schemas/createBillRequest'
        description: Bill that needs to be generated
        required: true
      responses:
        "201":
          content:
            application/vnd.worldpay.bill-v1+json:
              schema:
                $ref: '#/components/schemas/billDetails'
          description: A JSON returning the created bill.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
          description: Invalid Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
          description: Internal Server Error
      summary: Create a new Bill
      tags:
      - Bill
  /bills/{billId}:
    get:
      description: Returns a single Bill
      operationId: getBillById
      parameters:
      - description: Bill ID to be provided
        explode: false
        in: path
        name: billId
        required: true
        schema:
          $ref: '#/components/schemas/billNumberReference'
        style: simple
      - explode: false
        in: header
        name: X-WP-User-Id
        required: false
        schema:
          $ref: '#/components/schemas/userReference'
        style: simple
      responses:
        "200":
          content:
            application/vnd.worldpay.bill-v1+json:
              schema:
                $ref: '#/components/schemas/billDetails'
          description: successful operation
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
          description: Invalid Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
          description: Resource Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
          description: Internal Server Error
      summary: Find bill by billId
      tags:
      - Bill
    put:
      operationId: updateBill
      parameters:
      - description: Bill ID to be provided
        explode: false
        in: path
        name: billId
        required: true
        schema:
          $ref: '#/components/schemas/billNumberReference'
        style: simple
      - explode: false
        in: header
        name: X-WP-User-Id
        required: false
        schema:
          $ref: '#/components/schemas/userReference'
        style: simple
      requestBody:
        content:
          application/vnd.worldpay.bill-v1+json:
            schema:
              $ref: '#/components/schemas/updateBillRequest'
        description: Bill that needs to be updated
        required: true
      responses:
        "200":
          content:
            application/vnd.worldpay.bill-v1+json:
              schema:
                $ref: '#/components/schemas/billDetails'
          description: A JSON returning the updated Bill.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
          description: Invalid Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
          description: Resource Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
          description: Internal Server Error
      summary: Update an existing Bill by billId
      tags:
      - Bill
  /users/{userId}:
    get:
      description: Returns a user details
      operationId: getUserById
      parameters:
      - explode: false
        in: path
        name: userId
        required: true
        schema:
          $ref: '#/components/schemas/userReference'
        style: simple
      responses:
        "200":
          content:
            application/vnd.worldpay.bill-v1+json:
              schema:
                $ref: '#/components/schemas/user'
          description: successful operation
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
          description: Invalid Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
          description: Resource Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
          description: Internal Server Error
      summary: Verify user ID
      tags:
      - Bill
components:
...
  schemas:
    createBillRequest:
      properties:
        billTag:
          description: |
            The bill tag reference defined by the merchant to uniquely identify the bill with reference to tag say - table or room or invoice.
          example: TABLE-11 or ROOM-101
          maxLength: 30
          minLength: 1
          pattern: "^[a-zA-Z0-9_-]{1,30}$"
          type: string
        userId:
          description: |
            The userId defined by the merchant to uniquely identify  user. Must contain only alpha numeric, hypen (-) and underscore (_) characters.
          example: user_123-A54B3DF
          maxLength: 30
          minLength: 1
          pattern: "^[a-zA-Z0-9_-]{1,30}$"
          type: string
        totalAmount:
          allOf:
          - $ref: '#/components/schemas/amount'
          description: |
            The total amount of the Bill in minor currency units. A bill with a negative represents a refund.
          minimum: -999999999
      required:
      - userId
      type: object
    updateBillRequest:
      properties:
        userId:
          description: |
            The userId defined by the merchant to uniquely identify  user. Must contain only alpha numeric, hypen (-) and underscore (_) characters.
          example: user_123-A54B3DF
          maxLength: 30
          minLength: 1
          pattern: "^[a-zA-Z0-9_-]{1,30}$"
          type: string
        status:
          description: |
            All possible values for bill.

            `pending` - once bill is open to be processed

            `in-progress` - used to update once bill is picked-up for payment.

            `complete` - once bill is completely paid

            `partial-complete` - if bill is partially paid(in split payment) and wants to pay to the counter OR customer wants to update the bill in un-happy path

            `cancel` - to cancel a bill
          enum:
          - in-progress
          - partial-complete
          - pending
          - complete
          - cancel
          type: string
        paymentDetails:
          items:
            $ref: '#/components/schemas/paymentDetails'
          type: array
      required:
      - status
      - userId
      type: object
    billDetails:
      example:
        billTag: ""
        totalAmount: ""
        itemizedBill: itemizedBill
        billId: ""
        billDateTime: ""
        paymentDetails:
        - receiptCustomerDeclaration: receiptCustomerDeclaration
          merchantTransactionReference: MS_123-A54B3DF
          transactionDateTime: 2000-01-23T04:56:07.000+00:00
          taxFreeVoucher: MDEKMDJWT1VDSEVSCjEyCjA2RVhBTVBMRSBDT05URU5UCjEz
          merchant: null
          retrievalReferenceNumber: 6027456182
          paymentInstrument: null
          paymentType: null
          result: null
          paypoint:
            terminalId: "987654321"
            paypointId: ""
          gatewayTransactionReference: 1234567890AB
          eftSequenceNumber: 1
          value:
            amount: 1099
            dcc:
              convertedAmount: 1099
              convertedCurrencyCode: GBP
              conversionRate: 1.11647
            cashbackAmount: 1099
            currencyCode: GBP
            donationAmount: 1099
            gratuityAmount: 1099
          receiptRetentionReminder: receiptRetentionReminder
          receiptNumber: 1465812981
        - receiptCustomerDeclaration: receiptCustomerDeclaration
          merchantTransactionReference: MS_123-A54B3DF
          transactionDateTime: 2000-01-23T04:56:07.000+00:00
          taxFreeVoucher: MDEKMDJWT1VDSEVSCjEyCjA2RVhBTVBMRSBDT05URU5UCjEz
          merchant: null
          retrievalReferenceNumber: 6027456182
          paymentInstrument: null
          paymentType: null
          result: null
          paypoint:
            terminalId: "987654321"
            paypointId: ""
          gatewayTransactionReference: 1234567890AB
          eftSequenceNumber: 1
          value:
            amount: 1099
            dcc:
              convertedAmount: 1099
              convertedCurrencyCode: GBP
              conversionRate: 1.11647
            cashbackAmount: 1099
            currencyCode: GBP
            donationAmount: 1099
            gratuityAmount: 1099
          receiptRetentionReminder: receiptRetentionReminder
          receiptNumber: 1465812981
        userId: ""
        status: null
      properties:
        billId:
          allOf:
          - $ref: '#/components/schemas/billNumberReference'
          description: |
            A unique Id associated with this bill
          readOnly: true
        billTag:
          allOf:
          - $ref: '#/components/schemas/billTagReference'
          description: |
            A Tag associated with the given BillId. e.g. Table/Room
          readOnly: true
        totalAmount:
          allOf:
          - $ref: '#/components/schemas/amount'
          description: |
            The total amount of the Bill in minor currency units. A bill with a negative represents a refund.
          minimum: -999999999
          readOnly: true
        userId:
          allOf:
          - $ref: '#/components/schemas/userReference'
          description: |
            User who is assigned to this bill.
          readOnly: true
        billDateTime:
          allOf:
          - $ref: '#/components/schemas/billDateTime'
          description: |
            The most recent timestamp associated with bill.
          readOnly: true
        status:
          $ref: '#/components/schemas/billStatus'
        itemizedBill:
          description: |
            An itemized bill from Order management.
          maxLength: 10000
          minLength: 0
          pattern: "^[?@+'\"`~^%&\\*\\[\\]\\{\\}.!#|\\\\\\($¢€£₽₹¥₿฿元);,:;=\\/\\(\\\
            ),\\-\\w\\s+]*$"
          readOnly: true
          type: string
        paymentDetails:
          items:
            $ref: '#/components/schemas/paymentDetails'
          type: array
      required:
      - status
      type: object
    user:
      example:
        userName: userName
        userId: userId
      properties:
        userId:
          description: user id
          type: string
        userName:
          description: user name
          type: string
      type: object
...
    billNumberReference:
      description: |
        The bill number reference defined by the merchant to uniquely identify the bill.
      example: OR_123-A54B3DF
      maxLength: 30
      minLength: 1
      pattern: "^[a-zA-Z0-9_-]{1,30}$"
      type: string
    billTagReference:
      description: |
        The bill tag reference defined by the merchant to uniquely identify the bill with reference to tag say - table or room or invoice.
      example: TABLE-11 or ROOM-101
      maxLength: 30
      minLength: 1
      pattern: "^[a-zA-Z0-9_-]{1,30}$"
      type: string
...
    amount:
      description: |
        An amount of money in minor currency units.
      example: 1099
      format: int64
      maximum: 999999999
      minimum: 0
      type: integer
...
    errors:
      allOf:
      - $ref: '#/components/schemas/error'
      properties:
        validationErrors:
          description: A list of validation errors
          items:
            $ref: '#/components/schemas/validationError'
          type: array
      type: object
    error:
      properties:
        errorName:
          description: A machine and human readable error type for clarity and semantic
            understanding of the error.
          type: string
        message:
          description: A human readable message giving a corrective action for the
            error. This is not for machine consumption.
          type: string
      required:
      - errorName
      - message
      type: object
    validationError:
      allOf:
      - $ref: '#/components/schemas/error'
      properties:
        jsonPath:
          description: This field represents the JSONPath  of the element within the
            request body associated with the error
          type: string
        pathParameter:
          description: This field identifies a path parameter within the request URL
            that caused an error - e.g. a resource reference
          type: string
        queryParameter:
          description: This field represents the name of a query parameter within
            the request URL that caused the error
          type: string
      type: object
    userReference:
      description: |
        The userId defined by the merchant to uniquely identify  user. Must contain only alpha numeric, hypen (-) and underscore (_) characters.
      example: user_123-A54B3DF
      maxLength: 30
      minLength: 1
      pattern: "^[a-zA-Z0-9_-]{1,30}$"
      type: string
    billDateTime:
      description: |
        The most recent date-time asociated with the bill. Represented as a date-time according to RFC 3339, section 5.6, including the relevant time zone.
      format: date-time
      type: string
    billStatus:
      description: "All possible values for bill.\n\n`ready` - used to get the bill\
        \ from order management \n\n`pending` - once bill is open to be processed\n\
        \n`in-progress` - used to update once bill is picked-up for payment.\n\n`complete`\
        \ - once bill is completely paid\n\n`partial-complete` - if bill is partially\
        \ paid(in split payment) and wants to pay to the counter OR customer wants\
        \ to update the bill in un-happy path\n\n`cancel` - to cancel a bill\n"
      enum:
      - in-progress
      - partial-complete
      - pending
      - complete
      - cancel
      - ready
      type: string
    listBills_billStatus_parameter_inner:
      default: pending
      oneOf:
      - $ref: '#/components/schemas/billStatus'
      type: string
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http